Variable buffer resolution + fix warnings

+++ ojgl/src/utility/Timepoint.cpp
@@ -25,7 +25,7 @@ Timepoint Timepoint::now()
 #ifdef _DEBUG
     auto duration = std::chrono::system_clock::now().time_since_epoch();
     auto millis = std::chrono::duration_cast<std::chrono::milliseconds>(duration).count();
-    return Timepoint(millis);
+    return Timepoint(static_cast<long>(millis));
 #else
     return Timepoint(GetTickCount());
 #endif
